Refactor to split the large LV component
