+++ lib/elixir_console_web/live/console_live/history_component.ex
@@ -1,0 +1,7 @@
+defmodule ElixirConsoleWeb.ConsoleLive.HistoryComponent do
+  use Phoenix.LiveComponent
+
+  def render(assigns) do
+    Phoenix.View.render(ElixirConsoleWeb.ConsoleView, "history.html", assigns)
+  end
+end
