+++ lib/elixir_console_web/live/console_live/sidebar_component.ex
@@ -1,0 +1,7 @@
+defmodule ElixirConsoleWeb.ConsoleLive.SidebarComponent do
+  use Phoenix.LiveComponent
+
+  def render(assigns) do
+    Phoenix.View.render(ElixirConsoleWeb.ConsoleView, "sidebar.html", assigns)
+  end
+end
