+++ lib/elixir_console_web/templates/console/history.html.leex
@@ -1,0 +1,10 @@
+<div class="flex-1"></div>
+  <div class="p-2" id="commandOutput" phx-update="append">
+    <%= for output <- @output do %>
+      <div id="command<%= output.id %>" class="text-gray-300 font-medium"><%= print_prompt() %><%= format_command(output.command) %></div>
+      <div id="output<%= output.id %>" class="text-teal-300">
+        <%= if output.result do output.result end %>
+        <%= if output.error do %><span class="text-pink-400"><%= output.error %></span><% end %>
+      </div>
+    <% end %>
+</div>
